+++ dom/cache/CacheStorageParent.h
@@ -22,7 +22,7 @@ public:
   CacheStorageParent(const nsACString& aOrigin);
   virtual ~CacheStorageParent();
   virtual void ActorDestroy(ActorDestroyReason aReason) MOZ_OVERRIDE;
-  virtual bool RecvCreate(const uint32_t& aRequestId, const nsString& aKey) MOZ_OVERRIDE;
+  virtual bool RecvCreate(const uint64_t& aRequestId, const nsString& aKey) MOZ_OVERRIDE;
 private:
   nsRefPtr<CacheManager> mCacheManager;
 };
